Refactor plans modal with own css
