+++ client/my-sites/earn/memberships/products.jsx
@@ -26,7 +26,7 @@ import {
 	getSelectedSiteId,
 	getSelectedSiteSlug,
 } from 'calypso/state/ui/selectors';
-import RecurringPaymentsPlanAddEditModal from './add-edit-plan-modal';
+import RecurringPaymentsPlanAddEditModal from '../components/add-edit-plan-modal';
 import { ADD_NEW_PAYMENT_PLAN_HASH, ADD_NEWSLETTER_PAYMENT_PLAN_HASH } from './constants';
 import RecurringPaymentsPlanDeleteModal from './delete-plan-modal';
 import MembershipsSection from './';
